Use sync mixin in auto class generation
